Changes to hooking into the system
