+++ src/edu/pugetsound/mathcs/nlp/architecture_nlp/features/TextAnalyzer.java
@@ -16,6 +16,7 @@ import edu.pugetsound.mathcs.nlp.util.PathFormat;
 import edu.pugetsound.mathcs.nlp.architecture_nlp.brain.DialogueActTag;
 import edu.pugetsound.mathcs.nlp.architecture_nlp.datag.DAClassifier;
 import edu.pugetsound.mathcs.nlp.architecture_nlp.features.stanford.StanfordSuite;
+import edu.pugetsound.mathcs.nlp.interact.SPFSemanticAnalyzer;
 import edu.pugetsound.mathcs.nlp.kb.KBController;
 import edu.stanford.nlp.ling.CoreAnnotations.NamedEntityTagAnnotation;
 import edu.stanford.nlp.ling.CoreAnnotations.PartOfSpeechAnnotation;
@@ -75,7 +76,8 @@ public class TextAnalyzer {
 	 */
 	public TextAnalyzer(KBController kb){
 		nlpAnalyzer = new StanfordSuite();		
-		semAnalyzer = new CFGSemanticAnalyzer(kb);
+		//semAnalyzer = new CFGSemanticAnalyzer(kb);
+		semAnalyzer = new SPFSemanticAnalyzer();
 		anaphoraAnalyzer = new AnaphoraAnalyzer();
 		standardizedForms = new HashMap<String, String>();
 		greetClose = new HashMap<String, DialogueActTag>();
